[AX12] Retourne une position correcte

+++ arduino/petitrobot/driver/controle_ax12.h
@@ -12,8 +12,8 @@
 extern AX12 motor[NB_MOTEURS];
 
 void cherche_moteurs(void);
-byte tourner(int id, int id_moteur, int position);
-byte get_position(int id, int id_moteur);
-byte lire(int id, int ordre, int id_moteur);
+int tourner(int id, int id_moteur, int position);
+int get_position(int id, int id_moteur);
+int lire(int id, int ordre, int id_moteur);
 
 #endif /* C_AX12_H */
